Add stub program
This is based on the ARMCM0-STM32F051-DISCOVERY demo. The Makefile has
had a bunch of cruft removed.

The .gdbinit and openocd.cfg were written for openocd and the ST-Link
present on the STM32F0DISCOVERY board.

Semihosting can be enabled by passing USE_SEMIHOSTING=yes to the make
command line after doing a clean.

Signed-off-by: Joel Stanley <joel@jms.id.au>